MB-59515 Fix Query with Too Big Response
On Query page, if a query result is too big
to be processed in the browser, the UI is stuck until the POST
request finishes.
This fix checks progressively if the output
gets too long, by using reportProgress property on the observable.
Since Angular promises don't support reportProgress property or
something similar, the request with the user query must be
done using an observable.
When the intermediate loaded output is bigger than
the set limit (250MB), the query is canceled and
the following message is displayed to the user,
without the "Show Anyway" option:
"The statement results are larger than the
 workbench limit of 250MB and cannot be displayed.
 Please use another client product to access them."
Fixes CBSE CBSE-15526.

Change-Id: Iae2a296ec0af8fe0b55367037d0dd4d9112c5360
Reviewed-on: https://review.couchbase.org/c/query-ui/+/201577
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>