MB-60650 - prevent user from entering very large values for `offset`
The KV documents API supports fetching batches of documents given only
a bucket, limit, and offset. Thus the user could ask for 20 documents
starting from the 100th document in `travel-sample`.

A customer reported server problems, and the logs revealed that the
user was fetching batches of documents via the UI with a very high
offset (> 1,449,896). It was decided to limit offset values to 1000 or
less when using the KV REST API.

Note that the UI can use *either* the KV API or the query service to
fetch groups of documents, and the query service doesn't cause the
same server problems with high values of offset. This fix only limits
values of offset when using the KV api.

Change-Id: I5acc207a9858ba2fc5733fd394caa530b7f94bf7
Reviewed-on: https://review.couchbase.org/c/query-ui/+/206333
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Well-Formed: Restriction Checker