MB-60432 - Error when creating advisor-suggested indexes
This was a regression between 7.2 and 7.6 caused by changes to support
advice and query plans for subqueries (MB-52360). That change allowed
subquery advice to be displayed, but it broke the "Create" button for
all indexes.

To test, running the index advisor on the following query:

  SELECT t1.city FROM `travel-sample`.inventory.landmark t1
  where country = "France";

Without this change, clicking the "Create" buttons will cause an error
message to appear. With this change, the index is created correctly.

Change-Id: Ieacb11df91995c655a74105f8e2f9f3d153ed3de
Reviewed-on: https://review.couchbase.org/c/query-ui/+/207590
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>