MB-61452 - fix race condition when running batch queries
This bug reported an issue when running batch queries, where queries
after the first one were not being run, and the results were not even
being updated for the first query.

The problem is easily visible when non-trivial queries are as part of a batch. For paste the following example into the query workbench and run it:

 select count(*) from `travel-sample` where foo = "bar";
 select count(*) from `travel-sample` where foo = "bar";
 select count(*) from `travel-sample` where foo = "bar";
 select count(*) from `travel-sample` where foo = "bar";
 select count(*) from `travel-sample` where foo = "bar";
 select count(*) from `travel-sample` where foo = "bar";

Instead of showing 6 results, it only shows one result, and the remaining queries are not run.

The problem here originated with the fix for MB-59515. That fixed
changes query execution from Promises to Observables, to ensure that a
query response doesn't download so much result data that the browser
runs out of memory. However, batch queries relied on Promises to
indicated when one query is complete and the next is ready to run.

The fix is to create a special Promise for the query results, and only
resolve/reject that Promise when the Observable finishes.

Change-Id: I96d2b8a877f4cf60d961251872f0ebbe01f11ef8
Reviewed-on: https://review.couchbase.org/c/query-ui/+/208541
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Well-Formed: Restriction Checker