MB-61819 - allow document UI to show documents with null character in ID
While the UI does not support creating documents with null characters
in the document ID, KV and the REST API support it. Up to now this was
causing a failure in the document viewer. A null character is a valid
part of a string in javascript, however it appeared to cause HTML
rendering issues when rendering the generated HTML of the document
viewer table.

The solution is, when rendering, to replace the null character with
something else. In this case, '\0', which is a traditional rendering
of the null character.

See the Jira ticket for information about how to create one of these documents.

Change-Id: I97c4c2835eb233987315822c5b52f50737dad79b
Reviewed-on: https://review.couchbase.org/c/query-ui/+/209734
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>