MB-62439 - UI losing precision on numbers between 2^53 and 2^63
A customer reported large numbers (between 2^53 and 2^63) losing
precision when displayed in the UI. In general, javascript loses
precision when dealing with numbers >= 2^53. The UI had code to handle
those cases by intercepting the query response and converting large
numbers to strings. However, a recent change (to avoid browser crashes
with very large query responses > 250GiB) caused the intercept to be
ignored.

This fix restores the call to convert large numbers to strings, so
they do not lose precision in the UI.

The following query is a good test - in the previous UI precision is
lost for the number 2^53 + 1 and above. With the new UI, precision is
only lost for 2^63 and above (and that is due to the query service).

SELECT x,
       `2^53 - 1`,
       `2^53 + 1`,
       `2^63 - 1`,
       `2^63`
LET x = 393939393939393,
    `2^53 - 1` = 9007199254740991,
    `2^53 + 1` = 9007199254740993,
    `2^63 - 1` = 9223372036854775807,
    `2^63` = 9223372036854775808;

Change-Id: I3e0d2ac30b369a7869fb4e1bf7de615b59ab619a
Reviewed-on: https://review.couchbase.org/c/query-ui/+/213384
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Well-Formed: Restriction Checker