MB-61437 - problem when navigating to bucket with no indexes
The Documents page disables the WHERE textbox when the current bucket
has no indexes. However, it remembers the WHERE textbox contents
between visits to the Documents page. This presents a problem when:

- visiting the Documents page for a bucket with indexes, and setting a
  value for WHERE

- navigating to the Buckets page, selecting a bucket with no indexes,
  and clicking the link for the Documents page

The previous value for WHERE is still present, but the user can't
remove it because the text field is disabled.

The solution is to clear the WHERE field any time the user navigates
to the Documents page specifying a new Bucket.

Change-Id: I584f931ca2ac13f4b2d8852479263c5e250e9928
Reviewed-on: https://review.couchbase.org/c/query-ui/+/219670
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>