MB-64583 - Browser-based import failing with large documents
In CBSE-18799, the customer reported that browser-based import was
failing when trying to import a file with ~7MB documents. The UI was
designed to not attempt to import documents larger than 1MB, and to
show an error message in that case.

There are two related issues in the UI code causing this behavior:

1) The UI (implemented in Angular 8) was using an unfinished Angular 8
version of the Alert service. Instead it should have used the upgraded
AngularJS alert service. This change uses the correct service.

2) The 1MB limit is not necessary with the current query service. This
change increases the limit to 10MB, which works in testing.

Also, the warning message shown when loading a large file is expanded
to suggest using `cbimport` for better performance with large files.

Change-Id: Id2593a2660942a445a923356a6a535817b9e288e
Reviewed-on: https://review.couchbase.org/c/query-ui/+/220850
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>