MB-64042 - Add UI support for inline JavaScript user-defined functions
Until now, the UI has supported two types of UDFs:

- inline, where the function is defined as a string of sql++, and

- library javascript, where a javascript library is uploaded, and the
  function is defined as the name of one of the functions from that
  library.

It turns out that the query service has always supported a third type
of UDF, inline javascript, where there is no external library, just a
string of javascript. However, if the user created such a UDF from
SQL++, it would appear in the UI and not be editable.

This change adds UI support for inline javascript UDFs.

Change-Id: Ice7b1d2d2b1c2e97a0ea3bd6754621f1cf2d48ad
Reviewed-on: https://review.couchbase.org/c/query-ui/+/221760
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>