MB-65420 - check collection-level permissions in doc editor
The bug noted that the document editor was checking bucket-level
permissions but not collection-level permissions. Thus it was allowing
users to try and edit documents when the edit operation would fail.

The fix is as follows: when retrieving the list of scopes/collections
for a bucket, also hit `checkPermissions` to see the
read/upsert/delete privileges for each collection. Don't show any
collection in the menu for which the user doesn't have at least `read`
permission, and only enable edit and delete when the user has upsert
or delete privileges.

Change-Id: I90b0b15cb85e1695dda0161b65b91aeae52da415
Reviewed-on: https://review.couchbase.org/c/query-ui/+/224283
Tested-by: Eben Haber <eben@couchbase.com>
Tested-by: Piotr Najda  <piotr.najda@couchbase.com>
Reviewed-by: Piotr Najda  <piotr.najda@couchbase.com>