MB-67524 - dont wait for ADVISE before enabling "Execute" button
This issue was reported as part of CBSE-20377. In cases of complicated
queries and/or data, the ADVICE command can sometimes take
significantly longer than running the query itself. The problem for
the customer is that the query appears to take much longer than it
did, because the "Cancel" button remains active.

The problem is that the Query UI runs `EXPLAIN` and `ADVISE` with
every query, and doesn't re-enable the `Execute` button until all
three are done. `EXPLAIN` has never been a problem, since it takes
10ms or less in most cases. `ADVISE`, however, can sometimes take a
very long time.

This problem was already identified and fixed in the Capella UI. The
solution is to not wait for `ADVISE` to complete.

Change-Id: Ib6e2c5d514af2f8ecefcbab77801312423b0809f
Reviewed-on: https://review.couchbase.org/c/query-ui/+/230377
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>