MB-67993 - Don't run `explain` on inappropriate queries.
Check out the CBSE for an example of the problematic behavior:

https://jira.issues.couchbase.com/browse/CBSE-20816

A video of the error and fix can be seen here:

https://drive.google.com/file/d/1tLIYu-Dq6JGByccR_4ZzTN_CN0y6MhIM/view?usp=sharing

The gist is as follows: the query workbench automatically runs EXPLAIN
with every query since it is a cheap operation and having the query
plan can help the user understand why the query ran fast or
slow. However, we shoud not run EXPLAIN for DDL queries, transaction
queries, or execution of prepared queries.

The initial code still ran EXPLAIN for queries when it should
not. This change improves the filter.

Change-Id: Ia21588aea8a86e9bc4fe5e7f6e0efda7a3901a9e
Reviewed-on: https://review.couchbase.org/c/query-ui/+/232002
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Restriction Checker