MB-68089 - 'use_cbo' option not sent for internal statements
The query workbench sends two kinds of queries to the server:

- user queries, where the user has typed out the query in the workbench

- internal queries, run in support of UI functionality, such as
  fetching metadata or running EXPLAIN queries in the background when
  the user runs a regular query.

The problem identified in this ticket was that the 'use_cbo' option
was not being included with internal queries, meaning that background
EXPLAIN queries didn't accurately describe how a user's query was
actually run.

The solution is simple - include the 'use_cbo' option for internal as
well as user queries.

Change-Id: Icc43ec397337ad85823758104d82de9078a34855
Reviewed-on: https://review.couchbase.org/c/query-ui/+/232603
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>