MB-68775 - Query options dialog silently unavailable to non-Admin users
The fix for MB-66272 changed the `Use CBO` query option from boolean
(on/off) to three-value (on, off, or use the default value from the
query service settings). The default `useCBO` value was fetched from
the `/settings/querySettings` endpoint, and it turns out that the
endpoint is not readable by non-Admin users.

As a result, when a non-Admin user tries to open the query options
dialog, the code tries to retrieve the current settings and fails. An
error should have been surfaced to the user, but the code was using an
obsolete function for showing errors, so there was only an exception
in the browser console.

This change includes the following fixes: 1) if the call to
`/querySettings` fails, return a dummy object where `useCBO` is
undefined. 2) in the query options dialog, mention that the query
service `useCBO` option is 'unknown'. 3) fix the issue of the obsolete
error reporting function in qw.query.component.js, so that errors are
shown to the user.

Change-Id: I836d939c46df441e01b5d2b1c560a2797e40e6f7
Reviewed-on: https://review.couchbase.org/c/query-ui/+/234513
Reviewed-by: Eben Haber <eben@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>