MB-68775 - Query options dialog silently unavailable to non-Admin users
The fix for MB 66272 changes the `Use CBO` query option from boolean
to three-value (on, off, or use the default value from the query
service settings). As an aide to the user, the current value of the
setting from the Query Service was shown in the menu. However, it
turns out that the `/settings/querySettings` endpoint is not readable
by non-Admin users.

As a result, when a non-Admin user tries to open the query options
dialog, the code tries to retrieve the current settings and fails. An
error should have been surfaced to the user, but the code was using an
obsolete function for showing errors, so there was only an exception
in the browser console.

This change includes the following fixes:

1) add a check for rbac.cluster.settings.read

2) if no permissions to read settings, mark useCBO as undefined

3) if useCBO is undefined, don't display a value in the menu

4) fix the issue of the obsolete error reporting function in
   qw.query.component.js, so that any possible errors are shown to
   the user.

Change-Id: Ifc2c615467f7ddca82f12f1bc979e0027ec4cf89
Reviewed-on: https://review.couchbase.org/c/query-ui/+/234624
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>