MB-68949 - History import broken in the UI
This issue probably originated in the AngularJS -> Angular 8
conversion of the query workbench.

An obsolete line of code referring to a non-existent field was causing
query history import to fail with a undefined exception. That line of
code was removed in this change.

The error wasn't surfaced because the code was referring to an
obsolete function for bringing up an error dialog. This change fixes
that.

Also, drag-and-drop of text files to load the Query Editor was broken
due to a typo, referring to qc.handleDragOver and qc.handleFileDrop,
when in fact the `qc` was not needed.

Change-Id: I04cb38ec652c5240c1f8b273f3f9b1d088abd042
Reviewed-on: https://review.couchbase.org/c/query-ui/+/235173
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Restriction Checker